+++ src/modules/browse/search/interfaces/SearchBarProps.tsx
@@ -1,0 +1,6 @@
+import DietCategory from "../../types/DietCategory";
+
+export default interface SerchBarProps {
+  updateDietNameSearch(newDietNameSearch: string): void;
+  updateDietCategoriesSearch(newDietCategoriesSearch: DietCategory[]): void;
+}
